enable test pins only on dev

+++ app/Http/Controllers/Cron/CronController.php
@@ -30,14 +30,16 @@ class CronController extends Controller
      */
     public function enableTestPins(Request $request)
     {
+        if(env('APP_ENV') != 'live' && strpos("blitzerapp", $_SERVER['HTTP_HOST']) === false) {
+            Artisan::call("migrate:reset");
+            Artisan::call("migrate");
+            Artisan::call("db:seed");
+            
+            date_default_timezone_set(isset($request->timezone) ? $request->timezone : "Europe/Zagreb");
+            DB::table((new Pin)->getTable())
+                ->update(['updated_at' => date("Y-m-d H:i:s")]);
+            echo "Done. Check the app!";        
+        }
 
-        Artisan::call("migrate:reset");
-        Artisan::call("migrate");
-        Artisan::call("db:seed");
-        
-        date_default_timezone_set(isset($request->timezone) ? $request->timezone : "Europe/Zagreb");
-        DB::table((new Pin)->getTable())
-            ->update(['updated_at' => date("Y-m-d H:i:s")]);
-        echo "Done. Check the app!";
     }
 }
